feat: fixes urls and file names
