+++ packages/liferay-npm-scripts/src/jest/mocks/Headers.js
