+++ packages/liferay-npm-scripts/src/jest/setup.js
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-global.Headers = require('./mocks/headers');
+global.Headers = require('./mocks/Headers');
 
-global.Liferay = require('./mocks/liferay');
+global.Liferay = require('./mocks/Liferay');
 
 global.fetch = require('jest-fetch-mock');
